+++ CHANGELOG.md
@@ -26,6 +26,7 @@ to [Semantic Versioning](http://semver.org/).
 
 - The `Web` API:
     - `chance.tld()`
+    - `chance.domain()`
 
 - `unique_array()` assertion for the test suite.